Rework
I’ve made a lot of changes to the colorization, the interactivity etc.
The eye is now one set of colors each wake cycle, a wake cycle can be
reset, and once the eye is closed, a new eye colorization will be
generated either via a random time or a mouse click.